swampdragon_notifications/requirements.txt
swampdragon
django-model-utils
